refactor(db): squash alembic versions & update docs and tests
