+++ tests/test_cli_smoke.py
@@ -33,7 +33,7 @@ def test_cli_db_history():
 
     assert result.exit_code == 0, result.output
     revisions = {r["revision"] for r in json.loads(result.output)}
-    assert {"0001_gwapi_initial", "0002_log_tables"} <= revisions
+    assert revisions == {"0001_gwapi_initial"}
 
 
 def test_cli_db_current():
